Run publishing of `base-types`
