+++ tests/buildSrc/src/main/kotlin/io/spine/internal/gradle/Publish.kt
@@ -111,13 +111,13 @@ class Publish : Plugin<Project> {
         val extension = PublishExtension.create(project)
         project.extensions.add(PublishExtension::class.java, extensionName, extension)
 
-        project.afterEvaluate {
-            val soloMode = extension.soloProject != null
-            val rootPublish: Task? =
-                if (soloMode) null
-                else project.createPublishTask()
-            val checkCredentials = project.createCheckTask(extension)
+        val soloMode = extension.singleProject()
+        val rootPublish: Task? =
+            if (soloMode) null
+            else project.createPublishTask()
+        val checkCredentials: Task = project.createCheckTask(extension)
 
+        project.afterEvaluate {
             if (soloMode) {
                 project.applyMavenPublish(extension, null, checkCredentials)
             } else {
@@ -322,7 +322,7 @@ private constructor(
      *
      * If set, [projectsToPublish] will be ignored.
      */
-    var soloProject: Project? = null
+    private var soloProject: Project? = null
 
     init {
         spinePrefix.convention(true)
@@ -332,15 +332,17 @@ private constructor(
      * Instructs to publish the passed project _instead_ of [projectsToPublish].
      */
     fun publish(project: Project) {
-        soloProject = project;
+        soloProject = project
     }
+
+    fun singleProject(): Boolean = soloProject != null
 }
 
 /**
  * Configures the `spinePublishing` extension.
  *
  * As `Publish` is a class-plugin in `buildSrc`, we don't get strongly typed generated helper
- * methods for the `spinePublishing` configuration. Thus, we proviude this helper function for use
+ * methods for the `spinePublishing` configuration. Thus, we provide this helper function for use
  * in Kotlin build scripts.
  */
 @Suppress("unused")
